pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "JeremyLP", email = "jeremyleprunenec.pro@gmail.com"},
]
dependencies = [
    "Pillow>=9.4.0",
    "moviepy>=1.0.3",
    "setuptools>=67.5.0",
    "click>=8.1.3",
    "PyYAML>=6.0",
    "jsonschema>=4.17.3",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "python picgenius"
test = "pytest"
